+++ shared/atan/actor/mailbox/priority_mailbox.h
@@ -61,7 +61,7 @@ public:
         return this->queue_.empty();
     }
 
-    unsigned long size() override {
+    size_t size() override {
         return this->queue_.size();
     }
