+++ shared/atan/actor/routing/smallest_mailbox_pool.cpp
@@ -4,7 +4,7 @@ void smallest_mailbox_pool::tell_one(std::unique_ptr<message> msg) {
     std::lock_guard<std::mutex> lock(this->mailbox_mutex_);
 
     int actor_number = 0;
-    unsigned long smallest_mailbox_size = 0;
+    size_t smallest_mailbox_size = 0;
 
     for(int i = 0; i < this->routees_.size(); i++) {
         if(smallest_mailbox_size >= routees_[i]->mailbox_size()) {
